requirements.txt
alembic==1.5.2
aniso8601==8.1.0
asgiref==3.3.1
attrs==20.3.0
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
Django==3.1.7
djangorestframework==3.12.2
djangorestframework-simplejwt==4.6.0
idna==2.10
iniconfig==1.1.1
itsdangerous==1.1.0
Jinja2==2.11.2
lxml==4.6.2
Mako==1.1.4
MarkupSafe==1.1.1
marshmallow==3.9.1
marshmallow-sqlalchemy==0.24.1
mysql-connector==2.2.9
mysql-connector-python==8.0.22
packaging==20.8
passlib==1.7.4
pluggy==0.13.1
protobuf==3.14.0
psycopg2-binary==2.8.6
py==1.10.0
pyaes==1.6.1
pyasn1==0.4.8
PyJWT==2.0.1
pyparsing==2.4.7
pytest==6.2.1
python-dateutil==2.8.1
python-editor==1.0.4
pytz==2020.4
requests==2.25.1
rsa==4.6
six==1.15.0
SQLAlchemy==1.3.20
sqlparse==0.4.1
Telethon==1.18.2
toml==0.10.2
urllib3==1.26.2
Werkzeug==1.0.1
